+++ spec/selectors/article_spec.rb
@@ -2,99 +2,52 @@
 
 describe "article selector" do
   describe "locator" do
-    context "when <article> element" do
-      context "when direct descendant of body" do
-        it "finds the first element" do
-          render <<~HTML
-            <article>Content</article>
-          HTML
+    context "with <article> element" do
+      it "finds an article" do
+        render <<~HTML
+          <article data-test-id="test">Content</article>
+          <div>Content</div>
+        HTML
 
-          expect(page).to have_selector :article, count: 1
-        end
+        expect(page.find(:article)).to eq page.find(:test_id, "test")
+      end
 
-        it "finds based on [aria-label]" do
-          render <<~HTML
-            <article aria-label="Article article">Content</article>
-          HTML
+      it "finds based on part accessible name" do
+        render <<~HTML
+          <article aria-label="Article one" data-test-id="test">Content</article>
+          <article aria-label="Article two">Content</article>
+        HTML
 
-          expect(page).to have_selector :article, "Article article"
-        end
+        expect(page.find(:article, "Article o")).to eq page.find(:test_id, "test")
+      end
 
-        it "does not find differing on [aria-label]" do
-          render <<~HTML
-            <article aria-label="Article article">Content</article>
-          HTML
+      it "finds based on exact accessible name" do
+        render <<~HTML
+          <article aria-label="Article" data-test-id="test">Content</article>
+          <article aria-label="Article two">Content</article>
+        HTML
 
-          expect(page).to have_no_selector :article, "Not the right locator"
-        end
+        expect(page.find(:article, "Article", exact: true)).to eq page.find(:test_id, "test")
+      end
 
-        it "finds based on [aria-labelledby]" do
-          render <<~HTML
-            <article aria-labelledby="article_label">
-              <h1 id="article_label">Article article</h1>
-            </article>
-          HTML
+      it "finds based on regular expression" do
+        render <<~HTML
+          <article aria-label="Article one" data-test-id="test">Content</article>
+          <article aria-label="Article two">Content</article>
+        HTML
 
-          expect(page).to have_selector :article, "Article article"
-        end
-
-        it "does not find differing on [aria-labelledby]" do
-          render <<~HTML
-            <article aria-labelledby="article_label">
-              <h1 id="article_label">Article article</h1>
-            </article>
-          HTML
-
-          expect(page).to have_no_selector :article, "Not the right locator"
-        end
+        expect(page.find(:article, /Article o/)).to eq page.find(:test_id, "test")
       end
     end
 
-    context "when [role=article]" do
-      context "when direct descendant of body" do
-        it "finds the first element" do
-          render <<~HTML
-            <div role="article">Content</div>
-          HTML
+    context "with [role=article]" do
+      it "finds an article" do
+        render <<~HTML
+          <div role="article" data-test-id="test">Content</article>
+          <div>Content</div>
+        HTML
 
-          expect(page).to have_selector :article, count: 1
-        end
-
-        it "finds based on [aria-label]" do
-          render <<~HTML
-            <div role="article" aria-label="Div article">Content</div>
-          HTML
-
-          expect(page).to have_selector :article, "Div article"
-        end
-
-        it "does not find differing on [aria-label]" do
-          render <<~HTML
-            <div role="article" aria-label="Div article">Content</div>
-          HTML
-
-          expect(page).to have_no_selector :article, "Not the right locator"
-        end
-
-        it "finds based on [aria-labelledby]" do
-          render <<~HTML
-            <div role="article" aria-labelledby="article_label">
-              <h1 id="article_label">Div article</h1>
-            </div>
-          HTML
-
-          expect(page).to have_selector :article, "Div article"
-        end
-
-        it "does not find differing on [aria-labelledby]" do
-          render <<~HTML
-            <div role="article" aria-labelledby="article_label">
-              <h1 id="article_label">Div article</h1>
-            </div>
-          HTML
-
-          expect(page).to have_no_selector :article, "Not the right locator"
-        end
+        expect(page.find(:article)).to eq page.find(:test_id, "test")
       end
     end
   end